code for functional bookchapter wrt newUI
