+++ renderer/src/layouts/TopMenuBar.js
@@ -43,7 +43,7 @@ export default function TopMenuBar() {
                   <div className="hidden md:block">
                     <div className="-ml-4 flex items-baseline space-x-4">
                       <span className="text-white px-3 py-2 text-lg tracking-wide font-bold uppercase">
-                        Autographa
+                        <a href="/home">Autographa</a>
                         <span className="text-primary font-extrabold"> 2.0</span>
                       </span>
                     </div>
